Move contents of src/chase/chase.rs to src/chase.rs
