+++ benches/perf_test.rs
@@ -2,13 +2,13 @@ use rusty_razor::formula::syntax::Theory;
 use rusty_razor::chase::selector::Linear;
 use rusty_razor::chase::strategy::FIFO;
 use rusty_razor::chase::bounder::DomainSize;
-use rusty_razor::chase::chase::StrategyNode;
-use rusty_razor::chase::chase::Strategy;
-use rusty_razor::chase::chase::Selector;
+use rusty_razor::chase::StrategyNode;
+use rusty_razor::chase::Strategy;
+use rusty_razor::chase::Selector;
 use rusty_razor::chase::r#impl::basic::BasicEvaluator;
 use rusty_razor::chase::r#impl::basic::BasicSequent;
 use rusty_razor::chase::r#impl::basic::BasicModel;
-use rusty_razor::chase::chase::solve_all;
+use rusty_razor::chase::solve_all;
 use criterion::Criterion;
 use criterion::criterion_group;
 use criterion::criterion_main;
